CtrlVQE

Documentation for CtrlVQE.

CtrlVQE.FullyTrotterized

CtrlVQE.Systematic

CtrlVQE.Systematic

CtrlVQE.UniformWindowed

CtrlVQE.FullyTrotterized — Method

FullyTrotterized(signal::Signals.AbstractSignal, T::Real, r::Int)

Break a signal up so that each time-step is parameterized separately.

Usually you'll want to use this with constant signals.

CtrlVQE.Systematic — Function

Systematic(DeviceType, n, pulses; kwargs...)

Standardized constructor for a somewhat realistic DeviceType, but of arbitrary size.

Arguments

DeviceType::Type{<:Devices.Device} - the type of the device to be constructed

n::Int - the number of qubits in the device

pulses - a vector of control signals (Signals.AbstractSignal), or one to be copied

Unless otherwise stated, a systematic device has one channel for each qubit.

CtrlVQE.Systematic — Method

Systematic(TransmonDeviceType, n, pulses; kwargs...)

Standardized constructor for a transmon device.

This is a linearly coupled device, with uniformly-spaced resonance frequencies, and with all coupling and anharmonicity constants equal for each qubit. The actual values of each constant are meant to roughly approximate a typical IBM device.

Keyword Arguments

m::Int - the number of transmon levels to include in simulations

F::Type{<:AbstractFloat} - the precision type to use for device parameters

CtrlVQE.UniformWindowed — Method

UniformWindowed(signal::Signals.AbstractSignal, T::Real, W::Int)

Break a signal up into equal-sized windows.

Usually you'll want to use this with constant signals.