Bases

CtrlVQE.Bases — Module

Bases

Enumerates various linear-algebraic bases for representing statevectors and matrices.

CtrlVQE.Bases.Coordinate — Type

Coordinate(), aka COORDINATE

The eigenbasis of local quadrature operators $Q ≡ (a + a')/√2$.

CtrlVQE.Bases.Dressed — Type

Dressed(), aka DRESSED

The eigenbasis of the static Hamiltonian associated with a Device. Eigenvectors are ordered to maximize similarity with an identity matrix. Phases are fixed so that the diagonal is real.

CtrlVQE.Bases.Momentum — Type

Momentum(), aka MOMENTUM

The eigenbasis of local quadrature operators $P ≡ i(a - a')/√2$.

CtrlVQE.Bases.Occupation — Type

Occupation(), aka OCCUPATION

The eigenbasis of local number operators $n̂ ≡ a'a$. Generally equivalent to what is called the "Z basis", or the "computational basis".