Cost Functions

CtrlVQE.CostFunctions — Module

CostFunctions

Interfaces time-evolution and penalty functions to interface directly with optimization.

Each distinct cost function is implemented in a sub-module, alongside its gradient function. The sub-module implements a functions method which constructs a cost function and its corresponding gradient function. Thus, all you need to run a gradient based optimization is:

import CostFunctions.MySubModule: functions
f, g = functions(args...)

CtrlVQE.CostFunctions.AbstractCostFunction — Type

AbstractCostFunction

Super-type for "cost functions", to be plugged directly into optimization algorithms.

Implementation

Any concrete sub-type CF must be a callable object, accepting a vector of parameters. That is, it must implement the function (f::CF)(x̄::AbstractVector).

CtrlVQE.CostFunctions.AbstractCostFunction — Method

(f::AbstractCostFunction)(x̄::AbstractVector)

A cost function evaluated at the point in parameter space given by x̄.

CtrlVQE.CostFunctions.AbstractGradientFunction — Type

AbstractGradientFunction

Super-type for "gradient functions", to be plugged directly into optimization algorithms.

Each gradient function is implicitly associated with a cost function.

Implementation

Any concrete sub-type GF must be a callable object, accepting a vector of partial derivatives (to be calculated and filled in), and a vector of parameters. That is, it must implement the function (g::GF)(∇f̄::AbstractVector,x̄::AbstractVector).

CtrlVQE.CostFunctions.AbstractGradientFunction — Method

(g::AbstractGradientFunction)(∇f̄::AbstractVector, x̄::AbstractVector)

The gradient of a function at the point in parameter space given by x̄.

The result is stored in the first argument ∇f̄, as well as returned.

CtrlVQE.CostFunctions.AbstractGradientFunction — Method

(g::AbstractGradientFunction)(x̄::AbstractVector)

The gradient of a function at the point in parameter space given by x̄.

CtrlVQE.CostFunctions.CompositeCostFunction — Type

CompositeCostFunction(f̄::Vector{AbstractCostFunction})

The sum of several other cost-functions.

Use this eg. to combine an energy function with one or more penalty functions.

CtrlVQE.CostFunctions.CompositeCostFunction — Method

CompositeCostFunction(f̄...)

Alternate constructor, letting each function be passed as its own argument.

CtrlVQE.CostFunctions.CompositeGradientFunction — Type

CompositeGradientFunction(ḡ::Vector{AbstractCostFunction})

The sum of several other gradient-functions.

CtrlVQE.CostFunctions.CompositeGradientFunction — Method

CompositeGradientFunction(ḡ...)

Alternate constructor, letting each function be passed as its own argument.

Energy Functions

CtrlVQE.CostFunctions.EnergyFunctions.BareEnergy.functions — Method

functions(O0, ψ0, T, device, r; kwargs...)

Cost and gradient functions for the expectation value of a Hermitian observable.

The frame rotation (if provided) is applied to the molecular hamiltonian, rather than to the state.

Arguments

O0: a Hermitian matrix, living in the physical Hilbert space of device.

ψ0: the reference state, living in the physical Hilbert space of device.

T::Real: the total time for the state to evolve under the device Hamiltonian.

device::Devices.Device: the device

r::Int: the number of time steps to calculate the gradient signal

Keyword Arguments

algorithm::Evolutions.Algorithm: which algorithm to evolve ψ0 with. Defaults to Evolutions.rotate(r).

basis::Bases.BasisType: which basis O0 and ψ0 are represented in. ALSO determines the basis in which time-evolution is carried out. Defaults to Bases.OCCUPATION.

frame::Operators.StaticOperator: which frame to measure expecation values in. Use Operators.STATIC for the drive frame, which preserves the reference energy for a zero pulse. Use Operators.UNCOUPLED for the interaction frame, a (presumably) classically tractable approximation to the drive frame. Defaults to Operators.IDENTITY.

Returns

f: the cost function

g: the gradient function

CtrlVQE.CostFunctions.EnergyFunctions.ProjectedEnergy.functions — Method

functions(O0, ψ0, T, device, r; kwargs...)

Cost and gradient functions for the expectation value of a Hermitian observable.

The statevector is projected onto a binary logical space after time evolution, modeling an ideal quantum measurement where leakage is fully characterized.

The frame rotation (if provided) is applied to the molecular hamiltonian, rather than to the state.

Arguments

O0: a Hermitian matrix, living in the physical Hilbert space of device.

ψ0: the reference state, living in the physical Hilbert space of device.

T::Real: the total time for the state to evolve under the device Hamiltonian.

device::Devices.Device: the device

r::Int: the number of time steps to calculate the gradient signal

Keyword Arguments

algorithm::Evolutions.Algorithm: which algorithm to evolve ψ0 with. Defaults to Evolutions.rotate(r).

basis::Bases.BasisType: which basis O0 and ψ0 are represented in. ALSO determines the basis in which time-evolution is carried out. Defaults to Bases.OCCUPATION.

frame::Operators.StaticOperator: which frame to measure expecation values in. Use Operators.STATIC for the drive frame, which preserves the reference energy for a zero pulse. Use Operators.UNCOUPLED for the interaction frame, a (presumably) classically tractable approximation to the drive frame. Defaults to Operators.IDENTITY.

Returns

f: the cost function

g: the gradient function

CtrlVQE.CostFunctions.EnergyFunctions.NormalizedEnergy.functions — Method

functions(O0, ψ0, T, device, r; kwargs...)

Cost and gradient functions for the expectation value of a Hermitian observable.

The statevector is projected onto a binary logical space after time evolution, and then renormalized, modeling quantum measurement where leakage is completely obscured.

The frame rotation (if provided) is applied to the molecular hamiltonian, rather than to the state.

Arguments

O0: a Hermitian matrix, living in the physical Hilbert space of device.

ψ0: the reference state, living in the physical Hilbert space of device.

T::Real: the total time for the state to evolve under the device Hamiltonian.

device::Devices.Device: the device

r::Int: the number of time steps to calculate the gradient signal

Keyword Arguments

algorithm::Evolutions.Algorithm: which algorithm to evolve ψ0 with. Defaults to Evolutions.rotate(r).

basis::Bases.BasisType: which basis O0 and ψ0 are represented in. ALSO determines the basis in which time-evolution is carried out. Defaults to Bases.OCCUPATION.

frame::Operators.StaticOperator: which frame to measure expecation values in. Use Operators.STATIC for the drive frame, which preserves the reference energy for a zero pulse. Use Operators.UNCOUPLED for the interaction frame, a (presumably) classically tractable approximation to the drive frame. Defaults to Operators.IDENTITY.

Returns

f: the cost function

g: the gradient function

CtrlVQE.CostFunctions.EnergyFunctions.Normalization.functions — Method

functions(ψ0, T, device, r; kwargs...)

Cost and gradient functions for the norm of a statevector in a binary logical space.

Arguments

ψ0: the reference state, living in the physical Hilbert space of device.

T::Real: the total time for the state to evolve under the device Hamiltonian.

device::Devices.Device: the device

r::Int: the number of time steps to calculate the gradient signal

Keyword Arguments

algorithm::Evolutions.Algorithm: which algorithm to evolve ψ0 with. Defaults to Evolutions.rotate(r).

basis::Bases.BasisType: which basis O0 and ψ0 are represented in. ALSO determines the basis in which time-evolution is carried out. Defaults to Bases.OCCUPATION.

Returns

f: the cost function

g: the gradient function

Penalty Functions

CtrlVQE.CostFunctions.HardBounds.functions — Method

functions(λ̄, x̄L, x̄R, σ̄)

A steep exponential penalty for each parameter exceeding its bounds.

Arguments

λ̄: vector of weights for each penalty Set λ̄[i]=0 to skip penalties for the ith parameter.

x̄L: vector of lower bounds for each parameter

x̄R: vector of upper bounds for each parameter

σ̄: vector of scalings (smaller=steeper) for each penalty

Returns

f: the cost function

g: the gradient function

CtrlVQE.CostFunctions.SoftBounds.functions — Method

functions(λ̄, x̄L, x̄R, σ̄)

A shallow quadratic penalty for each parameter's deviation from mean.

Arguments

λ̄: vector of weights for each penalty Set λ̄[i]=0 to skip penalties for the ith parameter.

μ̄: vector of lower bounds for each parameter

σ̄: vector of scalings (smaller=steeper) for each penalty

Returns

f: the cost function

g: the gradient function

CtrlVQE.CostFunctions.SmoothBounds.functions — Method

functions(λ̄, x̄L, x̄R, σ̄)

A smooth exponential penalty for each parameter exceeding its bounds.

Arguments

λ̄: vector of weights for each penalty Set λ̄[i]=0 to skip penalties for the ith parameter.

μ̄R: vector of lower bounds for each parameter

μ̄L: vector of upper bounds for each parameter

σ̄: vector of scalings (smaller=steeper) for each penalty

Returns

f: the cost function

g: the gradient function