Evolutions

CtrlVQE.Evolutions — Module

Evolutions

Algorithms to run time evolution, and related constructs like gradient signals.

NOTE: the trapezoidaltimegrid function is a utility. But it is implicitly tied to evolutions. I'm not sure what to do about it. It is an awkward space, similar to QubitOperators below.

CtrlVQE.Evolutions.EvolutionType — Type

EvolutionType

Super-type for all evolution algorithms.

Implementation

Any concrete sub-type A must implement the following methods:

workbasis(::A): which Bases.BasisType the evolution algorithm uses

evolve!(::A, device, grid, ψ; callback=nothing): evolve ψ (in-place) on a time grid (you may assume the basis of ψ is the work basis)

You are allowed to implement evolve! for restricted types of grid (eg. require it to be a TrapezoidalIntegration), so long as you are clear in your documentation.

CtrlVQE.Evolutions.evolve — Function

evolve(evolution, device, [basis, ]grid, ψ0; result=nothing, kwargs...)

Evolve a state ψ0 over time grid under a device Hamiltonian, without mutating ψ0.

This method simply copies ψ0 (to result if provided, or else to a new array), then calls the mutating function evolve! on the copy. Please see evolve! for detailed documentation.

CtrlVQE.Evolutions.evolve! — Function

evolve!(evolution, device, [basis, ]T, ψ; basis=nothing, callback=nothing)

Evolve a state ψ by time T under a device Hamiltonian.

This method both mutates and returns ψ.

Arguments

evolution::EvolutionType: which evolution algorithm to use.

device::Devices.DeviceType: specifies which Hamiltonian to evolve under.

basis::Bases.BasisType: which basis ψ is represented in. Implicitly defaults to workbasis(evolution).

grid::TrapezoidalIntegration: defines the time integration bounds (eg. from 0 to T)

ψ: the initial statevector, defined on the full Hilbert space of the device.

Keyword Arguments

callback: a function which is called at each iteration of the time evolution. The function is passed three arguments: - i: indexes the iteration - t: the current time point - ψ: the current statevector, in the work basis The function is called after having evolved ψ into |ψ(t)⟩.

CtrlVQE.Evolutions.gradientsignals — Method

gradientsignals(device[, basis], grid, ψ0, r, O; kwargs...)

The gradient signals associated with a given device Hamiltonian, and an observable O.

Gradient signals are used to calculate analytical derivatives of a control pulse.

NOTE: Currently, this method assumes a trapezoidal rule, so only TrapezoidalIntegration grids are allowed.

Arguments

evolution::EvolutionType how to initialize the co-state |λ⟩ A standard choice would be ToggleEvolutions.Toggle(r).

device::Devices.DeviceType: specifies which Hamiltonian to evolve under. Also identifies each of the gradient operators used to calculate gradient signals.

basis::Bases.BasisType: which basis ψ is represented in. ALSO determines the basis in which calculations are carried out. Defaults to Bases.OCCUPATION.

grid::TrapezoidalIntegration: defines the time integration bounds (eg. from 0 to T)

ψ0: the initial statevector, defined on the full Hilbert space of the device.

O: a Hermitian observable, represented as a matrix. Gradients are calculated with respect to the expectation ⟨O⟩ at time T.

Keyword Arguments

result: an (optional) pre-allocated array to store gradient signals

callback: a function called at each iteration of the gradient signal calculation. The function is passed three arguments: - i: indexes the iteration - t: the current time point - ψ: the current statevector, in the OCCUPATION basis The function is called after having evolved ψ into |ψ(t)⟩, but before calculating ϕ̄[i,:]. Evolution here runs backwards.

Returns

A vector list ϕ̄, where each ϕ̄[:,j] is the gradient signal $ϕ_j(t)$ evaluated on the given time grid.

Explanation

A gradient signal $ϕ_j(t)$ is defined with respect to a gradient operator $Â_j$, an observable $Ô$, a time-dependent state |ψ(t)⟩, and total pulse duration T.

Let us define the expectation value $E(T) ≡ ⟨ψ(T)|Ô|ψ(T)⟩$.

Define the co-state $|λ(t)⟩$ as the (un-normalized) statevector which satisfies $E(T)=⟨λ(t)|ψ(t)⟩$ for any time t∊[0,T]. The gradient signal is defined as $ϕ_j(t) ≡ ⟨λ(t)|(iÂ_j)|ψ(t)⟩ + h.t.$.

gradientsignals(device[, basis], T, ψ0, r, Ō; kwargs...)

When the matrix argument O is replaced by a matrix list Ō, each Ō[:,:,k] represents a different Hermitian observable $Ô_k$. In this case, a different set of gradient signals is computed for each $Ô_k$.

Returns

A 3d array ϕ̄, where each ϕ̄[:,j,k] is the gradient signal $ϕ_j(t)$ defined with respect to the observable $Ô_k$.

Explanation

Multiple sets of gradient signals may be useful if you want to compute gradients with respect to multiple observables. For example, gradients with respect to a normalized molecular energy include contributions from both a molecular Hamiltonian and a leakage operator. This method enables such calculations using only a single "pass" through time.

CtrlVQE.Evolutions.workbasis — Method

workbasis(evolution::EvolutionType)

Which basis the evolution algorithm works in.

Also defines the default basis to interpret ψ as, in evolution methods.

Trotter Evolutions

CtrlVQE.ToggleEvolutions.Toggle — Type

Toggle

A Trotterization method (using r steps) alternately propagating static and drive terms.

The work basis for this algorithm is Bases.OCCUPATION, since the time-dependent "Drive" operator at each step is usually qubit-local.

NOTE: This method assumes a trapezoidal rule, so only TrapezoidalIntegration grids are allowed.

CtrlVQE.DirectEvolutions.Direct — Type

Direct

A Trotterization method (using r steps) calculating drive terms in the rotation-frame.

The work basis for this algorithm is Bases.DRESSED, since the rotating-frame evolution $U_t ≡ exp(-itH_0)$ happens at each step.

This algorithm exponentiates the matrix $U_t' V(t) U_t$ at each time step, so it is not terribly efficient.

NOTE: Currently, this method assumes a trapezoidal rule, so only TrapezoidalIntegration grids are allowed. TODO (mid): Actually I'm pretty sure this one doesn't have any reason to...