Integrations

CtrlVQE.Integrations — Module

Integrations

Encapsulations of everything you need to know to integrate over time.

CtrlVQE.Integrations.IntegrationType — Type

IntegrationType{F}

Super-type for all time grids, which encapsulate everything needed to integrate over time.

Type Parameters

F denotes the type for time values. Must be a real float.

Implementation

Any concrete sub-type G must implement the following methods:

nsteps(::G): the total number of finite jumps within this time grid.

This is the maximum index of the timeat and stepat functions, but it is NOT the length of the lattice, since the minimum index is 0.

timeat(::G, i::Int): time at index i (which may be zero)

stepat(::G, i::Int): stepsize at index i (which may be zero)

Let $r$ be the number of steps, $T$ be the duration, and $τ_i$ the step at index $i$. Then $∑_{i=0}^r τ_i = T$.

Base.reverse(::G): a new integration, identical except bounds are reversed

Base.reverse — Method

Base.reverse(grid::IntegrationType)

Construct a new integration, identical except bounds are reversed.

CtrlVQE.Integrations.duration — Method

duration(grid::IntegrationType)

The total duration, ie. the last time point minus the first.

CtrlVQE.Integrations.endtime — Method

endtime(grid::IntegrationType)

Upper bound of time integral.

CtrlVQE.Integrations.integrate — Function

integrate(grid::IntegrationType{F}, Φ)

Compute the integral $∫_0^T Φ(t)⋅dt$. Φ is a univariate scalar function of time, returning type F.

integrate(grid::IntegrationType, f̄::AbstractVector)

Treat the elements of f̄ as the function evaluations Φ(t) above. The length of f̄ must be nsteps(grid) + 1 and its eltype must be real.

integrate(grid::IntegrationType, Φ, f̄s::AbstractVector...)

Compute the integral $∫_0^T Φ(t, f1, f2...)⋅dt$, where f1 is the value of f̄s[1] at the index corresponding to time t, etc. Φ is a multivariate scalar function, of time and one argument for each f̄, returning type F. The length of each element in f̄s must be nsteps(grid) + 1.

CtrlVQE.Integrations.lattice — Method

lattice(grid::IntegrationType)

A vector of time points.

This is a regular vector, meaning it indexes from 1. This is offset by one from the timeat and stepat functions!

NOTE: This function is memoized, so DO NOT mutate its return value!

CtrlVQE.Integrations.nsteps — Method

nsteps(grid::IntegrationType)

The total number of finite jumps within this time grid.

This is the maximum index of the timeat and stepat functions, but it is NOT the length of the lattice, since the minimum index is 0.

CtrlVQE.Integrations.starttime — Method

starttime(grid::IntegrationType)

Lower bound of time integral.

CtrlVQE.Integrations.stepat — Method

stepat(grid::IntegrationType, i::Int)

The stepsize at index i (which may be zero).

Let $r$ be the number of steps, $T$ be the duration, and $τ_i$ the step at index $i$. Then $∑_{i=0}^r τ_i = T$.

CtrlVQE.Integrations.stepsize — Method

stepsize(grid::IntegrationType)

The average step size, ie. duration divided by number of steps.

CtrlVQE.Integrations.timeat — Method

timeat(grid::IntegrationType, i::Int)

The time at index i (which may be zero).

First-Order Integrations

CtrlVQE.TrapezoidalIntegrations.TrapezoidalIntegration — Type

TrapezoidalIntegration(t0::F, tf::F, r::Int)

First-order grid using a trapezoidal rule time integration.

Arguments

t0, tf: lower and upper bounds of the integral

r: the number of time steps (remember number of lattice points is r+1)

Explanation

This grid gives a first-order quadrature, meaning every step takes you forward in time.

Additionally, this grid uses uniform spacing. You'd think that would mean stepat(grid, i) would give $τ_i = τ$ for every i, where $τ ≡ T / r$ and $T ≡ tf - t0$. But careful! The sum of all τ_i must match the length of the integral, ie. T. But there are r+1 points, and (r+1)⋅τ > T. How do we reconcile this? A "Left Hand Sum" would omit t=T from the time points; a "Right Hand Sum" would omit t=0. The trapezoidal rule omits half a point from each.

That sounds awfully strange, but it's mathematically sound! We only integrate through half of each boundary time point t=0 and t=T. Thus, those points, and only those points, have a spacing of τ/2.