LinearAlgebraTools

CtrlVQE.LinearAlgebraTools — Module

LinearAlgebraTools

Implement some frequently-used linear-algebraic operations.

Much of this functionality is available in Julia's standard LinearAlgebra library, but the implementations here might take advantage of pre-allocated TempArrays and (sometimes) efficient tensor contractions.

CtrlVQE.LinearAlgebraTools.MatrixList — Type

MatrixList{T}

Semantic alias for a 3d array which explicitly represents a list of matrices.

Think of a MatrixList Ā as a list where each A[i] = Ā[:,:,i] is a matrix.

CtrlVQE.LinearAlgebraTools.VectorList — Type

VectorList{T}

Semantic alias for Matrix which explicitly represents a distinct vector in each column.

Think of a VectorList v̄ as a list where each v[i] = v̄[:,i] is a vector.

CtrlVQE.LinearAlgebraTools.basisvector — Method

basisvector(N::Int, i::Int)

A vector of N Bools, all zero except for index i, which is one.

CtrlVQE.LinearAlgebraTools.braket — Method

braket(x1::AbstractVector, A::AbstractMatrix, x2::AbstractVector)

Compute the braket ⟨x1|A|x2⟩.

CtrlVQE.LinearAlgebraTools.braket — Method

braket(x1::AbstractVector, ā::MatrixList, x2::AbstractVector)

Compute the braket ⟨x1|kron(ā)|x2⟩, but somewhat more efficiently.

CtrlVQE.LinearAlgebraTools.cis! — Function

cis!(A::AbstractMatrix, x=1)

Calculates $\exp(ixA)$ for a Hermitian matrix A.

The name comes from the identity exp(ix) = Cos(x) + I Sin(x).

Note that this method mutates A itself to the calculated exponential. Therefore, A must have a complex float type, and it must not be an immutable view. For example, even though A must be Hermitian for this method to work correctly, it can't actually be a LinearAlgebra.Hermitian view.

CtrlVQE.LinearAlgebraTools.cis_type — Method

cis_type(x)

Promote the number type of x to a complex float (compatible with cis operations).

The argument x may be a number, an array of numbers, or a number type itself.

CtrlVQE.LinearAlgebraTools.expectation — Method

expectation(ā::MatrixList, x::AbstractVector)

Compute the braket ⟨x|kron(ā)|x⟩, but somewhat more efficiently.

CtrlVQE.LinearAlgebraTools.expectation — Method

expectation(A::AbstractMatrix, x::AbstractVector)

Compute the braket $⟨x|A|x⟩$.

CtrlVQE.LinearAlgebraTools.kron — Method

kron(Ā::MatrixList; result=nothing)

The kronecker product of each matrix in Ā.

CtrlVQE.LinearAlgebraTools.kron — Method

kron(v̄::VectorList; result=nothing)

The kronecker product of each vector in v̄.

Ordering: [x1 x2] ⊗ [y1 y2] = [x1⋅y1 x1⋅y2 x2⋅y1 x2⋅y2]

Optionally, pass a pre-allocated array of compatible type and shape as result.

CtrlVQE.LinearAlgebraTools.rotate! — Function

rotate!(R, x)

Apply the rotation R to the object x, mutating x.

Generally, R is a unitary (or orthogonal) matrix. If x is a vector, rotate! computes $x ← Rx$. If x is a matrix, rotate! computes $x ← RxR'$.

You may also pass R as a MatrixList, which is interpreted as a rotation with a factorized tensor structure. In other words, if r̄ is a MatrixList, rotate!(r̄, x) is equivalent to rotate!(kron(r̄), x), except that the former has a more efficient implementation.

Since this method mutates x, the number type of x must be sufficiently expressive. For example, if R is a unitary matrix, x had better be a vector of complex floats.