Operators

CtrlVQE.Operators — Module

Operators

Enumerates various categories of Hermitian observable related to a device.

CtrlVQE.Operators.Channel — Type

Channel(i,t)

An individual drive term (indexed by i) at a specific time t.

For example, in a transmon device, Channel(q,t) might represent $Ω_q(t) [\exp(iν_qt) a_q + \exp(-iν_qt) a_q']$, the drive for a single qubit.

Note that you are free to have multiple channels for each qubit, or channels which operate on multiple qubits.

CtrlVQE.Operators.Coupling — Type

Coupling(), aka COUPLING

The components of the static Hamiltonian which are non-local to any one qubit.

For example, in a transmon device, Coupling() represents the sum $∑_{p,q} g_{pq} (a_p'a_q + a_q'a_p)$.

CtrlVQE.Operators.Drive — Type

Drive(t)

The sum of all drive terms at a specific time t.

This represents the sum of each Qubit(i), where i iterates over each drive term in the device.

For example, in a transmon device, Drive(t) might represent $∑_q Ω_q(t) [\exp(iν_qt) a_q + \exp(-iν_qt) a_q']$.

CtrlVQE.Operators.Gradient — Type

Gradient(j,t)

An individual gradient operator (indexed by j) at a specific time t.

The gradient operators appear in the derivation of each gradient signal, which are used to calculate analytical gradients of each variational parameter. The gradient operators are very closely related to individual channel operators, but sufficiently distinct that they need to be treated separately.

For example, for a transmon device, each channel operator $Ω_q(t) [\exp(iν_qt) a_q + \exp(-iν_qt) a_q']$ is associated with two gradient operators:

$\exp(iν_qt) a_q + \exp(-iν_qt) a_q'$

$i[\exp(iν_qt) a_q - \exp(-iν_qt) a_q']$

CtrlVQE.Operators.Hamiltonian — Type

Hamiltonian(t)

The full Hamiltonian at a specific time t.

This represents the sum of Static() and Drive(t).

CtrlVQE.Operators.Identity — Type

Identity(), aka IDENTITY

The identity operator.

CtrlVQE.Operators.Qubit — Type

Qubit(q)

The component of the static Hamiltonian which is local to qubit q.

For example, in a transmon device, Qubit(2) represents a term $ω_q a_q'a_q - δ_q/2~ a_q'a_q'a_q a_q$.

CtrlVQE.Operators.Static — Type

Static(), aka STATIC

All components of the static Hamiltonian.

This represents the sum of Uncoupled() and Coupled()

CtrlVQE.Operators.Uncoupled — Type

Uncoupled(), aka UNCOUPLED

The components of the static Hamiltonian which are local to each qubit.

This represents the sum of each Qubit(q), where q iterates over each qubit in the device.

For example, in a transmon device, Uncoupled() represents the sum $∑_q (ω_q a_q'a_q - δ_q/2~ a_q'a_q'a_q a_q)$.