Parameters

CtrlVQE.Parameters — Module

Parameters

Standardized interface for interacting with variational parameters.

Most commonly implemented by Signals and Devices.

CtrlVQE.Parameters.bind — Function

bind(entity, x̄::AbstractVector)

Assigns new values for each variational parameter in entity.

Implementation

This method should mutate entity such that, for example, the expression bind(entity, x̄); ȳ = values(entity); x̄ == ȳ evaluates true. There is no return value.

CtrlVQE.Parameters.count — Function

count(entity)

The number of variational parameters in entity.

Implementation

Must return a non-negative integer.

CtrlVQE.Parameters.names — Function

names(entity)

An ordered list of human-readable names for each variational parameter in entity.

Implementation

Must return a vector of strings.

CtrlVQE.Parameters.values — Function

values(entity)

An ordered list of the numerical values for each variational parameter in entity.

Implementation

Must return a vector of some float type.