QubitOperators

CtrlVQE.QubitOperators — Module

QubitOperators

Interfaces arbitrary physical Hilbert space with a strictly binary logical Hilbert space.

NOTE: I don't especially like how this module is organized, so consider this code subject to change.

CtrlVQE.QubitOperators.driveframe — Method

driveframe(H::AbstractMatrix, device::Devices.Device, T::Real)

Rewrite a molcular Hamiltonian in the drive frame of a device at time T.

This first projects H onto the physical Hilbert space, then evolves by the device's static Hamiltonian.

CtrlVQE.QubitOperators.interactionframe — Method

interactionframe(H::AbstractMatrix, device::Devices.Device, T::Real)

Rewrite a molcular Hamiltonian in the interaction frame of a device at time T.

This first projects H onto the physical Hilbert space, then evolves by the uncoupled parts of a device's static Hamiltonian. This is supposed to approximate the drive frame while remaining classically tractable.

CtrlVQE.QubitOperators.localqubitisometries — Method

localqubitprojectors(device::Devices.Device)

A matrix list of local qubit isometries for each individual qubit in the device.

CtrlVQE.QubitOperators.localqubitprojectors — Method

localqubitprojectors(device::Devices.Device)

A matrix list of local qubit projectors for each individual qubit in the device.

CtrlVQE.QubitOperators.nqubits — Method

nqubits(H)

Infer the number of qubits of a statevector or matrix living in a logical qubit space.

In other words, calculates n assuming the dimension of H is 2^n.

CtrlVQE.QubitOperators.project — Method

project(A, device::Devices.Device)

Extend a statevector or matrix living in a two-level space onto a physical Hilbert space.

CtrlVQE.QubitOperators.qubitisometry — Method

qubitisometry(device::Devices.Device)

An isometry from the physical Hilbert space of a device onto a logical two-level space.

If Φ is the isometry, and |ψ⟩ is a statevector living in the full Hilbert space, the vector Φ|ψ⟩ is a smaller vector living in the two-level space.

CtrlVQE.QubitOperators.qubitprojector — Method

qubitprojector(device::Devices.Device)

A projector from the physical Hilbert space of a device onto a logical two-level space.

The projector Π does not reduce the size of its operand; it only removes support on the non-logical states. If you want to change the size, use qubitisometry instead.

CtrlVQE.QubitOperators.reference — Method

reference(H)

The basis state which minimizes the matrix H.