Signals

CtrlVQE.Signals — Module

Signals

Time-dependent functions suitable for control signals with variational parameters.

The main motivation of this module is to provide a common interface for analytical gradients and optimization.

CtrlVQE.Signals.SignalType — Type

SignalType{P,R}

Super-type for all signal objects $Ω(t)$.

Type Parameters

P denotes the type of all variational parameters. Must be a real float.

R denotes the type of $Ω(t)$ itself. May be any number type.

Implementation

Any concrete sub-type S must implement all functions in the Parameters module.

In particular, Parameters.values(Ω::S) must return a vector of type P.

If you are trying to create your own signal type, you probably want to implement a ParametricSignal, which already has an implementation for the Parameters interface.

In addition, the following methods must be implemented:

valueat(Ω::S, t::Real): the actual function $Ω(t)$. Must return a number of type R.

partial(i::Int, Ω::S, t::Real): the partial derivative $∂Ω/∂x_i$ evaluated at time t, where $x_i$ is Ω's i-th variational parameter (ie. Parameters.names(Ω)[i]). Must return a number of type R.

Base.string(Ω::S, names::AbstractVector{String}): a human-readable description of the signal, inserting each element of names in the place of the corresponding parameter. For example, a complex constant signal might return a description like "A + i B", where A and B are the "names" given by the names argument.

CtrlVQE.Signals.SignalType — Method

(signal::SignalType{P,R})(t::Real)

Syntactic sugar: if Ω is a SignalType, then Ω(t) gives valueat(Ω,t).

Base.string — Method

Base.string(Ω::SignalType, names::AbstractVector{String})

Substitutes the default name of each variational parameter for the ones in names.

Base.string — Method

Base.string(Ω::SignalType)

A human-readable description of the signal. Returns type string.

CtrlVQE.Signals.partial — Method

partial(
    i::Int, signal::SignalType{P,R}, t̄::AbstractVector{<:Real};
    result=nothing,
)

Vectorized version. Returns a vector of type R.

Optionally, pass a pre-allocated array of compatible type and shape as result.

CtrlVQE.Signals.partial — Method

partial(i::Int, signal::SignalType{P,R}, t::Real)

The partial derivative $∂Ω/∂x_i|_t$. Returns a number of type R.

Here $x_i$ is the signal's i-th variational parameter (ie. Parameters.names(signal)[i]).

CtrlVQE.Signals.valueat — Method

valueat(signal::SignalType{P,R}, t̄::AbstractVector{<:Real}; result=nothing)

Vectorized version. Returns a vector of type R.

Optionally, pass a pre-allocated array of compatible type and shape as result.

CtrlVQE.Signals.valueat — Method

valueat(signal::SignalType{P,R}, t::Real)

The signal at time t, ie. $Ω(t)$. Returns a number of type R.

CtrlVQE.ParametricSignals.ConstrainedSignal — Type

ConstrainedSignal(constrained::<:ParametricSignal, constraints::Vector{Symbol})

The parametric signal constrained, freezing all fields in constraints.

Frozen parameters are omitted from the Parameters interface. In other words, they do not appear in Parameters.names or Parameters.values, and they are not mutated by Parameters.bind.

CtrlVQE.ParametricSignals.ConstrainedSignal — Method

ConstrainedSignal(constrained::ParametricSignal, constraints::Symbol...)

Alternate constructor, letting each field be passed as its own argument.

CtrlVQE.ParametricSignals.ParametricSignal — Type

ParametricSignal{P,R} ... end

Super-type for user-defined signal objects $Ω(t)$.

Type Parameters

P denotes the type of all variational parameters. Must be a real float.

R denotes the type of $Ω(t)$ itself. May be any number type.

Implementation

Any concrete sub-type S must be a mutable struct, and all of its variational parameters are of type P. By default, all fields of type P are treated as variational parameters. You may optionally change this by implementing parameters(::S), which should return a vector of each variational parameter of S. But before you do that, consider whether a ConstrainedSignal provides the desired behavior.

The following methods must be implemented:

Signals.valueat(Ω::S, t::Real): the actual function $Ω(t)$. Must return a number of type R.

Signals.partial(i::Int, Ω::S, t::Real): the partial derivative $∂Ω/∂x_i$ evaluated at time t, where $x_i$ is Ω's i-th variational parameter (ie. Parameters.names(Ω)[i]). Must return a number of type R.

Base.string(Ω::S, names::AbstractVector{String}): a human-readable description of the signal, inserting each element of names in the place of the corresponding parameter. For example, a complex constant signal might return a description like "A + i B", where A and B are the "names" given by the names argument.

CtrlVQE.ParametricSignals.parameters — Method

parameters(::S<:ParametricSignal{P,R})

A vector of all the variational parameters in a signal of type S.

CtrlVQE.ParametricSignals.parameters — Method

parameters(::Type{S<:ParametricSignal{P,R}})

A vector of all the variational parameters in a signal of type S.

CtrlVQE.WindowedSignals.WindowedSignal — Type

WindowedSignal(windows, starttimes)

A signal which applies a different function for each window.

Arguments

windows: a vector of signals

starttimes: a vector of times transitioning each window

Both windows and starttimes have the same length; starttimes[i] indicates when windows[i] begins.

This signal is undefined for times t < starttimes[1]. Normally, starttimes[1] == 0.

Note that each window must share the same type parameters P and R.

CtrlVQE.WindowedSignals.get_window_from_parameter — Method

get_window_from_parameter(signal::WindowedSignal, i::Int)

Identify the window index given a parameter index (by counting parameters in windows).

CtrlVQE.WindowedSignals.get_window_from_time — Method

get_window_from_time(signal::WindowedSignal, t::Real)

Identify the window index given the time (by inspecting starttimes).

CtrlVQE.CompositeSignals.CompositeSignal — Type

CompositeSignal(components::AbstractVector{<:SignalType{P,R}})

A signal which is the sum of each sub-signal in components.

Note that each component must share the same type parameters P and R.

CtrlVQE.CompositeSignals.CompositeSignal — Method

CompositeSignal(components::SignalType...)

Alternate constructor, letting each component be passed as its own argument.

CtrlVQE.CompositeSignals.WeightedCompositeSignal — Type

WeightedCompositeSignal(components::AbstractVector{<:SignalType{P,R}})

A signal which is the weighted average of each sub-signal in components.

Note that each component must share the same type parameters P and R.

CtrlVQE.CompositeSignals.WeightedCompositeSignal — Method

WeightedCompositeSignal(components::SignalType...)

Alternate constructor, letting each component be passed as its own argument.

Using this constructor, weights initialize to zero.

CtrlVQE.ModulatedSignals.ModulatedSignal — Type

ModulatedSignal(components::AbstractVector{<:SignalType{P,R}})

A signal which is the product of each sub-signal in components.

Note that each component must share the same type parameters P and R.

CtrlVQE.ModulatedSignals.ModulatedSignal — Method

ModulatedSignal(components::SignalType...)

Alternate constructor, letting each component be passed as its own argument.

Constant Signals

CtrlVQE.ConstantSignals.ComplexConstant — Type

ComplexConstant(A::F, B::F) where {F<:AbstractFloat}

The constant complex signal $Ω(t) = A + iB$.

CtrlVQE.ConstantSignals.Constant — Type

Constant(A::F) where {F<:AbstractFloat}

The constant real signal $Ω(t) = A$.

CtrlVQE.ConstantSignals.PolarComplexConstant — Type

PolarComplexConstant(r::F, ϕ::F) where {F<:AbstractFloat}

The constant complex signal $Ω(t) = r \exp(iϕ)$.

Interval Signals

CtrlVQE.IntervalSignals.ComplexInterval — Type

ComplexInterval(A::F, B::F, s1::F, s2::F) where {F<:AbstractFloat}

The piecewise signal $Ω(t) = A + iB$, for times t∈[s1,s2), and $Ω(t) = 0$ otherwise.

CtrlVQE.IntervalSignals.Interval — Type

Interval(A::F, s1::F, s2::F) where {F<:AbstractFloat}

The piecewise signal $Ω(t) = A$, for times t∈[s1,s2), and $Ω(t) = 0$ otherwise.

Step Function Signals

CtrlVQE.StepFunctionSignals.StepFunction — Type

StepFunction(A::F, s::F) where {F<:AbstractFloat}

The piecewise signal $Ω(t) = A⋅Θ(t-s)$, where $Θ$ is the Heaviside step function.

Gaussian Signals

CtrlVQE.GaussianSignals.Gaussian — Type

Gaussian(A::F, σ::F, s::F) where {F<:AbstractFloat}

A Gaussian real signal $Ω(t) = A \exp( -((t-s)/σ)^2 / 2)$.

Trigonometric Signals

CtrlVQE.TrigonometricSignals.Sine — Type

Sine(A::F, ν::F, ϕ::F) where {F<:AbstractFloat}

A real signal $Ω(t) = A \sin(νt+ϕ)$.

Sinc Signals

CtrlVQE.SincSignals.Sinc — Type

Sinc(A::F, σ::F, s::F) where {F<:AbstractFloat}

A real signal $Ω(t) = A \sin(χ)/χ$ where $χ≡(t-s)/σ$. Note that Julia's built-in sinc function implicitly sets A=1, s=0, and σ=1/π.

Tanh Signals

CtrlVQE.TanhSignals.Tanh — Type

Tanh(A::F, σ::F, s::F) where {F<:AbstractFloat}

A real signal $Ω(t) = A \sin(χ)/χ$ where $χ≡(t-s)/σ$. Note that Julia's built-in sinc function implicitly sets A=1, s=0, and σ=1/π.