CtrlVQE

Documentation for CtrlVQE.

Main Modules

Parameters

Bases

Operators

LinearAlgebraTools

Devices

Transmon Devices

Signals

Constant Signals

Interval Signals

Step Function Signals

Gaussian Signals

Evolutions

Trotter Evolutions

QubitOperators

Cost Functions

Energy Functions

Penalty Functions

Recipes

CtrlVQE.FullyTrotterized — Method

FullyTrotterized(signal::Signals.SignalType, T::Real, r::Int)

Break a signal up so that each time-step is parameterized separately.

Usually you'll want to use this with constant signals.

CtrlVQE.Systematic — Method

Systematic(TransmonDeviceType, n, pulses; kwargs...)

Standardized constructor for a somewhat realistic transmon device, but of arbitrary size.

This is a linearly coupled device, with uniformly-spaced resonance frequencies, and with all coupling and anharmonicity constants equal for each qubit. The actual values of each constant are meant to roughly approximate a typical IBM device.

Arguments

TransmonDeviceType: the type of the device to be constructed

n::Int: the number of qubits in the device

pulses: a vector of control signals (Signals.SignalType), or one to be copied

Keyword Arguments

m::Int: the number of transmon levels to include in simulations (defaults to 2)

F: the float type to use for device parameters (defaults to Float64)

CtrlVQE.UniformWindowed — Method

UniformWindowed(signal::Signals.SignalType, T::Real, W::Int)

Break a signal up into equal-sized windows.

Usually you'll want to use this with constant signals.

Miscellaneous Modules

Quples

CtrlVQE.Quples — Module

Quples

Qubit tuples: simple types to represent couplings within a device.

CtrlVQE.Quples.Quple — Type

Quple(q1,q2)

A (symmetric) coupling between qubits indexed by q1 and q2.

Note that the order is irrelevant: Quple(q1,q2) == Quple(q2,q1).

TempArrays

CtrlVQE.TempArrays — Module

TempArrays

Maintain caches for pre-allocated arrays used only temporarily within a function.

Do NOT use this module for arrays whose data are accessible outside the function in which they are created.

CtrlVQE.TempArrays.array — Function

array(::F, shape::Tuple, index=nothing)

Fetch a temporary array with type F and shape shape.

The index parameter is an additional unique key, allowing the module to cache mulitple arrays of the same type and shape. You should pass index=Symbol(@__MODULE__) to prevent collisions across modules. You may pass a tuple, eg. index=(Symbol(@__MODULE__), :otherkey) to prevent collisions within a module.

array(F::Type{<:Number}, shape::Tuple, index=nothing)

Same as above but passing the type directly, rather than an instance of the type.