Cost Functions

CtrlVQE.CostFunctions — Module

CostFunctions

Interfaces time-evolution and penalty functions to interface directly with optimization.

Each distinct cost function is implemented in a sub-module, alongside its gradient function. The sub-module implements a functions method which constructs a cost function and its corresponding gradient function. Thus, all you need to run a gradient based optimization is:

import CostFunctions.MySubModule: functions
f, g = functions(args...)

CtrlVQE.CostFunctions.CompositeCostFunction — Type

CompositeCostFunction(components::AbstractVector{CostFunctionType})

The sum of several cost-functions.

Use this eg. to combine an energy function with one or more penalty functions.

Note that all components must have matching eltype and length.

As a practical convenience, this struct maintains some dynamic attributes so that results for each component function can be inspected in the middle of an optimization. I don't really consider them an official part of the interface, but feel free to use them. Although...they aren't technically "thread-safe". ^_^

CtrlVQE.CostFunctions.CompositeCostFunction — Method

CompositeCostFunction(f̄::CostFunctionType{F}...)

Alternate constructor, letting each function be passed as its own argument.

CtrlVQE.CostFunctions.CostFunctionType — Type

CostFunctionType{F}

Super-type for "cost functions", to be plugged directly into optimization algorithms.

Implementation

Any concrete sub-type CF must implement the following methods:

Base.length(::CF): the number of parameters this cost function takes

cost_function(::CF): returns a Function which takes a parameter vector and returns the output of the cost function ie. a callable expression (x::Vector) -> f(x)

grad_function_inplace(::CF): returns a mutating Function which takes a gradient vector (to be mutated) and a parameter vector, and writes the gradient vector to the first argument. As a matter of habit, the resulting gradient vector should also be returned.

If your cost function involves calculating the expectation value of a time-evolved state, you should implement an EnergyFunction (even if it isn't strictly an energy). This type has a couple extra requirements to allow energy trajectories over the evolution.

CtrlVQE.CostFunctions.CostFunctionType — Method

(fn::CostFunctionType)(x̄::AbstractVector)

Evaluate the value of fn at the point x̄.

This is syntactic sugar for constructing a dedicated cost_function and calling it. It will not normally take advantage of cached work variables, so avoid using it in high-performance code.

CtrlVQE.CostFunctions.EnergyFunction — Type

CostFunctionType{F}

Super-type for "cost functions", to be plugged directly into optimization algorithms.

Implementation

Any concrete sub-type CF must implement everything required in the CostFunctionType interface, so consult the documentation for CostFunctionType carefully.

In additon, the following method must be implemented:

trajectory_callback(::CF, E::AbstractVector; callback=nothing) returns a Function compatible with Evolutions.evolve callback ie. a callable expression (i::Int, t::Real, ψ::Vector) -> Nothing which sets E[i] to the energy of a partially evolved wavefunction ψ. If callback is provided, the function calls that callback afterwards.

Finally, the following methods must now accept a keyword argument:

cost_function(::CF; callback=nothing): When callback is provided, the time evolution must call it at each timestep.

grad_function_inplace(::CF; ϕ=nothing): When ϕ is provided, write the gradient signals to it.

Base.eltype — Method

Base.eltype(fn::CostFunctionType)

Gives the number type for parameters of this cost function.

Base.length — Method

Base.length(fn::CostFunctionType)

Gives the number of parameters for this cost function.

CtrlVQE.CostFunctions.cost_function — Method

cost_function(fn::CostFunctionType)

Converts the struct fn into a literal Function of a parameter vector.

The function accepts a parameter vector x̄ which should have the type and length given by eltype(fn) and length(fn). The function returns the value of fn at the point x̄.

CtrlVQE.CostFunctions.cost_function — Method

cost_function(fn::CostFunctionType; callback=nothing)

Same as for CostFunctionType except that whenever the function is called, the time evolution calls callback (if provided) in each time step.

CtrlVQE.CostFunctions.grad_function — Method

grad_function(fn::CostFunctionType)

Constructs a Function to calculate the gradient of fn at a particular point.

The function accepts a parameter vector x̄ which should have the type and length given by eltype(fn) and length(fn). The function returns the gradient (a vector) of fn at the point x̄.

CtrlVQE.CostFunctions.grad_function_inplace — Method

grad_function_inplace(fn::CostFunctionType)

Constructs a (mutating) Function to calculate the gradient of fn at a particular point.

The function accepts a gradient vector ∇f̄ (to be mutated) and a parameter vector x̄. Both should have the type and length given by eltype(fn) and length(fn). After the function is called, ∇f̄ contains the gradient of fn at the point x̄.

CtrlVQE.CostFunctions.grad_function_inplace — Method

grad_function_inplace(fn::CostFunctionType; ϕ=nothing)

Same as for CostFunctionType except that whenever the function is called, ϕ (if provided) is updated to contain the gradient signals.

Energy Functions

CtrlVQE.BareEnergies.BareEnergy — Type

BareEnergy(O0, ψ0, T, device, r; kwargs...)

Expectation value of a Hermitian observable.

This type is called "bare" because it does not perform any projection steps (except perhaps what is hard-coded into the structure of the observable O0).

Arguments

evolution::Evolutions.TrotterEvolution: which algorithm to evolve ψ0 with A sensible choice is ToggleEvolutions.Toggle(r), where r is the number of Trotter steps. Must be a TrotterEvolution because the gradient signal is inherently Trotterized.

device::Devices.DeviceType: the device, which determines the time-evolution of ψ0

basis::Bases.BasisType: the measurement basis ALSO determines the basis which ψ0 and O0 are understood to be given in. An intuitive choice is Bases.OCCUPATION, aka. the qubits' Z basis. That said, there is some doubt whether, experimentally, projective measurement doesn't actually project on the device's eigenbasis, aka Bases.DRESSED. Note that you probably want to rotate ψ0 and O0 if you change this argument.

frame::Operators.StaticOperator: the measurement frame Think of this as a time-dependent basis rotation, which is applied to O0. A sensible choice is Operators.STATIC for the "drive frame", which ensures a zero pulse (no drive) system retains the same energy for any T. Alternatively, use Operators.UNCOUPLED for the interaction frame, a (presumably) classically tractable approximation to the drive frame, or Operators.IDENTITY to omit the time-dependent rotation entirely.

T::Real: the total time for the state to evolve under the device Hamiltonian.

ψ0: the reference state, living in the physical Hilbert space of device.

O0: a Hermitian matrix, living in the physical Hilbert space of device.

CtrlVQE.ProjectedEnergies.ProjectedEnergy — Type

ProjectedEnergy(O0, ψ0, T, device, r; kwargs...)

Expectation value of a Hermitian observable.

The statevector is projected onto a binary logical space after time evolution, modeling an ideal quantum measurement where leakage is fully characterized.

Arguments

evolution::Evolutions.TrotterEvolution: which algorithm to evolve ψ0 with A sensible choice is ToggleEvolutions.Toggle(r), where r is the number of Trotter steps. Must be a TrotterEvolution because the gradient signal is inherently Trotterized.

device::Devices.DeviceType: the device, which determines the time-evolution of ψ0

basis::Bases.BasisType: the measurement basis ALSO determines the basis which ψ0 and O0 are understood to be given in. An intuitive choice is Bases.OCCUPATION, aka. the qubits' Z basis. That said, there is some doubt whether, experimentally, projective measurement doesn't actually project on the device's eigenbasis, aka Bases.DRESSED. Note that you probably want to rotate ψ0 and O0 if you change this argument.

frame::Operators.StaticOperator: the measurement frame Think of this as a time-dependent basis rotation, which is applied to O0. A sensible choice is Operators.STATIC for the "drive frame", which ensures a zero pulse (no drive) system retains the same energy for any T. Alternatively, use Operators.UNCOUPLED for the interaction frame, a (presumably) classically tractable approximation to the drive frame, or Operators.IDENTITY to omit the time-dependent rotation entirely.

T::Real: the total time for the state to evolve under the device Hamiltonian.

ψ0: the reference state, living in the physical Hilbert space of device.

O0: a Hermitian matrix, living in the physical Hilbert space of device.

CtrlVQE.NormalizedEnergies.NormalizedEnergy — Type

NormalizedEnergy(O0, ψ0, T, device, r; kwargs...)

Expectation value of a Hermitian observable.

The statevector is projected onto a binary logical space after time evolution, and then renormalized, modeling quantum measurement where leakage is completely obscured.

Arguments

evolution::Evolutions.TrotterEvolution: which algorithm to evolve ψ0 with A sensible choice is ToggleEvolutions.Toggle(r), where r is the number of Trotter steps. Must be a TrotterEvolution because the gradient signal is inherently Trotterized.

device::Devices.DeviceType: the device, which determines the time-evolution of ψ0

basis::Bases.BasisType: the measurement basis ALSO determines the basis which ψ0 and O0 are understood to be given in. An intuitive choice is Bases.OCCUPATION, aka. the qubits' Z basis. That said, there is some doubt whether, experimentally, projective measurement doesn't actually project on the device's eigenbasis, aka Bases.DRESSED. Note that you probably want to rotate ψ0 and O0 if you change this argument.

frame::Operators.StaticOperator: the measurement frame Think of this as a time-dependent basis rotation, which is applied to O0. A sensible choice is Operators.STATIC for the "drive frame", which ensures a zero pulse (no drive) system retains the same energy for any T. Alternatively, use Operators.UNCOUPLED for the interaction frame, a (presumably) classically tractable approximation to the drive frame, or Operators.IDENTITY to omit the time-dependent rotation entirely.

T::Real: the total time for the state to evolve under the device Hamiltonian.

ψ0: the reference state, living in the physical Hilbert space of device.

O0: a Hermitian matrix, living in the physical Hilbert space of device.

CtrlVQE.Normalizations.Normalization — Type

Normalization(ψ0, T, device, r; kwargs...)

The norm of a statevector in a binary logical space.

Arguments

evolution::Evolutions.TrotterEvolution: which algorithm to evolve ψ0 with A sensible choice is ToggleEvolutions.Toggle(r), where r is the number of Trotter steps. Must be a TrotterEvolution because the gradient signal is inherently Trotterized.

device::Devices.DeviceType: the device, which determines the time-evolution of ψ0

basis::Bases.BasisType: the measurement basis ALSO determines the basis which ψ0 is understood to be given in. An intuitive choice is Bases.OCCUPATION, aka. the qubits' Z basis. That said, there is some doubt whether, experimentally, projective measurement doesn't actually project on the device's eigenbasis, aka Bases.DRESSED. Note that you probably want to rotate ψ0 if you change this argument.

T::Real: the total time for the state to evolve under the device Hamiltonian.

ψ0: the reference state, living in the physical Hilbert space of device.

Penalty Functions

CtrlVQE.HardBounds.HardBound — Type

HardBound(λ̄, x̄L, x̄R, σ̄)

A steep exponential penalty for each parameter exceeding its bounds.

Arguments

λ̄: vector of weights for each penalty Set λ̄[i]=0 to skip penalties for the ith parameter.

x̄L: vector of lower bounds for each parameter

x̄R: vector of upper bounds for each parameter

σ̄: vector of scalings (smaller=steeper) for each penalty

CtrlVQE.SoftBounds.SoftBound — Type

SoftBound(λ̄, μ̄, σ̄)

A shallow quadratic penalty for each parameter's deviation from mean.

Arguments

λ̄: vector of weights for each penalty Set λ̄[i]=0 to skip penalties for the ith parameter.

μ̄: vector of means for each parameter

σ̄: vector of scalings (smaller=steeper) for each penalty

CtrlVQE.SmoothBounds.SmoothBound — Type

SmoothBound(λ̄, x̄R, x̄L, σ̄)

A smooth exponential penalty for each parameter exceeding its bounds.

Arguments

λ̄: vector of weights for each penalty Set λ̄[i]=0 to skip penalties for the ith parameter.

x̄R: vector of upper bounds for each parameter

x̄L: vector of lower bounds for each parameter

σ̄: vector of scalings (smaller=steeper) for each penalty