Devices

CtrlVQE.Devices — Module

Devices

In silico representation of quantum devices, in which quantum states evolve in time.

In this package, the "static" components (ie. qubit frequencies, couplings, etc.) and the "drive" components (ie. control signal, variational parameters, etc.) are all integrated into a single DeviceType object. All you need to know how a quantum state ψ evolves up time T is in the device.

CtrlVQE.Devices.DeviceType — Type

DeviceType

Super-type for all device objects.

Implementation

Any concrete sub-type D must implement all functions in the Parameters module.

In particular, if any static operators in your device depend on variational parameters, you should consult the "Note on Caching" below.

In addition, all methods in the following sections must be implemented.

Counting methods

Algebra methods

Operator methods

Type methods

Gradient methods

If your device's drive channels are all local, you should implement a LocallyDrivenDevice, which has a few extra requirements.

Counting methods:

nqubits(::D): the number of qubits in the device - call this n.

nlevels(::D): the number of physical levels in each "qubit" - call this m.

ndrives(::D): the number of distinct drive channels.

ngrades(::D): the number of distinct gradient operators.

Each of these methods returns an integer.

Algebra methods:

localloweringoperator(::D): an m × m matrix applying the lowering operator a to a single qubit.

This method should define result=nothing as a keyword argument; when passed, use it as the array to store your result in.

Operator methods:

qubithamiltonian(::D, ā, q::Int): the static components of the device Hamiltonian local to qubit q.

staticcoupling(::D, ā): the static components of the device Hamiltonian nonlocal to any one qubit.

driveoperator(::D, ā, i::Int, t::Real): the distinct drive operator for channel i at time t

gradeoperator(::D, ā, j::Int, t::Real): the distinct gradient operator indexed by j at time t

Each of these methods should define result=nothing as a keyword argument; when passed, use it as the array to store your result in.

Each of these methods takes a 3darray ā; the annihilation operator $a_q$ is given by the matrix ā[:,:,q]. These methods should construct a new matrix as a function of each $a_q$. Usually, each ā[:,:,q] is defined on the full Hilbert space (ie. m^n × m^n), but sometimes the code exploits a simple tensor structure by passing in local m × m operators instead, so do not assume a specific size a priori.

The annihilation operators $a_q$ and their adjoints $a_q'$ form a complete algebra, so it is always possible to express any operator given just ā. For example, the Pauli spin matrices in a two-level system can be expressed as $X=a+a'$, $Y=i(a-a')$, and $Z=a'a$. If you really want to write your operators as functions of something other than $a_q$, you may "hack" in a new algebra by implementing new methods for this module's eltype_algebra, algebra, and localalgebra functions.

Type methods:

eltype_localloweringoperator(::D)

eltype_qubithamiltonian(::D)

eltype_staticcoupling(::D)

eltype_driveoperator(::D)

eltype_gradeoperator(::D)

Each of these methods gives the number type of the corresponding operator. Implement these methods based only on your implementation of the methods, ie. they should be independent of the type of ā.

Gradient methods:

gradient(::D, τ̄, t̄, ϕ̄): the gradient vector for each variational parameter in the device.

Each partial is generally an integral over at least one gradient signal. The arguments τ̄ and t̄ are time spacings and time grids, as given by Evolutions.trapezoidaltimegrid. The argument ϕ̄ is a 2d array; ϕ̄[:,:,j] contains the jth gradient signal $ϕ_j(t)$ evaluated at each point in t̄. An integral $∫f(ϕ_j(t), t)⋅dt$ is computed as sum( f.(ϕ̄[:,:,j], t̄) .* τ̄ ).

This method should define result=nothing as a keyword argument; when passed, use it as the array to store your result in.

Notes on Caching

This module uses the Memoization package to cache some arrays as they are calculated.

This does not apply to any method which depends on an absolute time t, though it does apply to methods depending only on a relative time τ. For example, the propagator for a static Hamiltonian is cached, but not one for a drive Hamiltonian. No caching happens for any method if its result keyword argument is used.

Usually, variational parameters only affect time-dependent methods, but if any of your device's static operators do depend on a variational parameter, you should be careful to empty the cache when Parameters.bind is called.

You can completely clear everything in the cache with:

Memoization.empty_all_caches!()

Alternatively, selectively clear caches for affected functions via:

Memoization.empty_cache!(fn)

I don't know if it's possible to selectively clear cached values for specific methods. If it can be done, it would require obtaining the actual Dict being used as a cache for a particular function, figuring out exactly how that cache is indexed, and manually removing elements matching your targeted method signature.

CtrlVQE.Devices.LocallyDrivenDevice — Type

LocallyDrivenDevice

Super-type for device objects whose drive channels act locally on individual qubits.

Inherit from this type if your driveoperator and gradeoperator methods depend only on a single annihilation operator ā[:,:,q]. This enables more efficient propagation methods which exploit a tensor product structure.

Implementation

Any concrete sub-type D must implement everything required in the DeviceType interface, so consult the documentation for DeviceType carefully.

In addition, the following methods must be implemented:

drivequbit(::D, i::Int): index of the qubit on which channel i is applied.

gradequbit(::D, j::Int): index of the qubit associated with the jth gradient operator.

It's usually trivial to infer the channel index i associated with each gradient operator, in which case gradequbit(device, j) = drivequbit(device, i), but this is left as an implementation detail.

Base.eltype — Method

Base.eltype(op::Operators.OperatorType, device::DeviceType[, basis::Bases.BasisType])

The number type of the matrix returned by operator(op, device, basis).

When omitted, the basis defaults to Bases.OCCUPATION.

CtrlVQE.Devices.algebra — Function

algebra(device::DeviceType[, basis::Bases.BasisType])

A matrix list ā, where ā[:,:,q] represents the annihilation operator $a_q$.

When omitted, the basis defaults to Bases.OCCUPATION.

Each ā[:,:,q] acts globally on the full Hilbert space of the device, even in bases where it acts trivially on states outside the local space of qubit q. To construct local operators, use localalgebra instead.

CtrlVQE.Devices.basisrotation — Method

basisrotation(tgt::Bases.BasisType, src::Bases.BasisType, device::DeviceType)

Calculate the basis rotation U which transforms $|ψ_{src}⟩ → |ψ_{tgt}⟩ = U|ψ_{src}⟩$.

CtrlVQE.Devices.basisrotation — Method

basisrotation(tgt::Bases.LocalBasis, src::Bases.LocalBasis, device::DeviceType, q::Int)

Same as above, except that U acts on the local Hilbert space of qubit q.

This is used elsewhere for more efficient rotations exploiting tensor structure.

CtrlVQE.Devices.braket — Method

braket(op, device[, basis], ψ1, ψ2)

The braket of an operator describing a device with respect to states ψ1 and ψ2.

If $A$ is the operator specified by op, this method calculates $⟨ψ1|A|ψ2⟩$.

Arguments

op::Operators.OperatorType: which operator to estimate (eg. static, drive, etc.).

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the state ψ is represented in. Defaults to Bases.OCCUPATION when omitted.

ψ1, ψ2: Statevectors defined over the full Hilbert space of the device.

CtrlVQE.Devices.diagonalize — Function

diagonalize(basis::Bases.BasisType, device::DeviceType)

Compute the vector of eigenvalues Λ and the rotation matrix U for a given basis.

U is an operator acting on the global Hilbert space of the device.

The result is packed into a LinearAlgebra.Eigen object, but it may be unpacked by Λ, U = diagonalize(basis, device).

diagonalize(basis::Bases.LocalBasis, device::DeviceType, q::Int)

Same as above, except that U acts on the local Hilbert space of qubit q.

Note that you can still construct the global rotation matrix of a local basis by using the first method.

CtrlVQE.Devices.driveoperator — Method

driveoperator(device::DeviceType, ā::MatrixList, i::Int, t::Real; result=nothing)

The distinct drive operator for channel i at time t.

This method is a function of annihilation operators $a_q$ given by ā[:,:,q], which are matrices acting globally on the physical Hilbert space. If device is a LocallyDrivenDevice, the matrices may also act on a local physical Hilbert space for each individual qubit.

Optionally, pass a pre-allocated array of compatible type and shape as result.

CtrlVQE.Devices.drivequbit — Method

drivequbit(device, i::Int)

Index of the qubit on which channel i is applied.

CtrlVQE.Devices.eltype_algebra — Function

eltype_algebra(device::DeviceType[, basis::Bases.BasisType])

The number type of each annihilation operator $a_j$ represented in the given basis.

When omitted, the basis defaults to Bases.OCCUPATION.

CtrlVQE.Devices.eltype_driveoperator — Method

eltype_driveoperator(device::DeviceType)

The number type of the time-dependent drive channels for this device.

The number type of the algebra ā is ignored for the purposes of this method.

CtrlVQE.Devices.eltype_gradeoperator — Method

eltype_gradeoperator(device::DeviceType)

The number type of the gradient operators for this device.

The number type of the algebra ā is ignored for the purposes of this method.

CtrlVQE.Devices.eltype_localloweringoperator — Method

eltype_localloweringoperator(device::DeviceType)

The number type of a local lowering operator for this device.

CtrlVQE.Devices.eltype_qubithamiltonian — Method

eltype_qubithamiltonian(device::DeviceType)

The number type of the local static components of the Hamiltonian for this device.

The number type of the algebra ā is ignored for the purposes of this method.

CtrlVQE.Devices.eltype_staticcoupling — Method

eltype_staticcoupling(device::DeviceType)

The number type of the non-local static components of the Hamiltonian for this device.

The number type of the algebra ā is ignored for the purposes of this method.

CtrlVQE.Devices.evolve! — Method

evolve!(op, device[, basis], t, ψ)

Propagate a state ψ by a time t under the Hermitian op describing a device.

This function is identical to propagate!, except that the cache is not used for intermediate propagator matrices, and that it is undefined for time-dependent operators. Look to the Evolutions module for algorithms compatible with time-dependence!

Arguments

op::Operators.OperatorType: which operator to evolve under (eg. static, drive, etc.).

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the state ψ is represented in. Defaults to Bases.OCCUPATION when omitted.

t::Real: the amount to move forward in time by.

ψ: Either a vector or a matrix, defined over the full Hilbert space of the device.

CtrlVQE.Devices.evolver — Method

evolver(op, device[, basis], t; kwargs...)

A unitary propagator describing evolution under a Hermitian operator for an time t.

This function is identical to propagator, except that the argument t is considered an absolute time so it is never cached, and that it is undefined for time-dependent operators. It exists solely to perform rotating-frame rotations at every time-step without worrying about over-caching.

Arguments

op::Operators.OperatorType: which operator to evolve under (eg. static, drive, etc.).

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the operator will be represented in. Defaults to Bases.OCCUPATION when omitted.

t::Real: the amount to move forward in time by.

Keyword Arguments

result: a pre-allocated array of compatible type and shape, used to store the result.

CtrlVQE.Devices.expectation — Method

expectation(op, device[, basis], ψ)

The expectation value of an operator describing a device with respect to the state ψ.

If $A$ is the operator specified by op, this method calculates $⟨ψ|A|ψ⟩$.

Arguments

op::Operators.OperatorType: which operator to estimate (eg. static, drive, etc.).

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the state ψ is represented in. Defaults to Bases.OCCUPATION when omitted.

ψ: A statevector defined over the full Hilbert space of the device.

CtrlVQE.Devices.globalize — Method

globalize(device::DeviceType, op::AbstractMatrix, q::Int; result=nothing)

Extend a local operator op acting on qubit q into the global Hilbert space.

Optionally, pass a pre-allocated array of compatible type and shape as result.

CtrlVQE.Devices.gradeoperator — Method

gradeoperator(device::DeviceType, ā::MatrixList, j::Int, t::Real; result=nothing)

The distinct gradient operator indexed by j at time t.

I have defined the "gradient operator" $Â_j$ as the Hermitian operator for which the jth gradient signal is $ϕ_j = ⟨λ|(iÂ_j)|ψ⟩ + h.t.$.

This method is a function of annihilation operators $a_q$ given by ā[:,:,q], which are matrices acting globally on the physical Hilbert space. If device is a LocallyDrivenDevice, the matrices may also act on a local physical Hilbert space for each individual qubit.

Optionally, pass a pre-allocated array of compatible type and shape as result.

CtrlVQE.Devices.gradequbit — Method

gradequbit(device, j::Int)

Index of the qubit associated with the jth gradient operator.

CtrlVQE.Devices.gradient — Method

gradient(::DeviceType, τ̄, t̄, ϕ̄; result=nothing)

The gradient vector of partials for each variational parameter in the device.

Each partial is generally an integral over at least one gradient signal. The arguments τ̄ and t̄ are time spacings and time grids, as given by Evolutions.trapezoidaltimegrid. The argument ϕ̄ is a 2d array; ϕ̄[:,:,j] contains the jth gradient signal $ϕ_j(t)$ evaluated at each point in t̄.

Optionally, pass a pre-allocated array of compatible type and shape as result.

CtrlVQE.Devices.localalgebra — Function

localalgebra(device::DeviceType[, basis::Bases.LocalBasis])

A matrix list ā, where ā[:,:,q] represents the annihilation operator $a_q$.

When omitted, the basis defaults to Bases.OCCUPATION.

Each ā[:,:,q] acts locally on the physical Hilbert space of qubit q. Note that you can construct global operators in a local basis by using algebra instead.

CtrlVQE.Devices.localbasisrotations — Method

localbasisrotations(tgt::Bases.LocalBasis, src::Bases.LocalBasis, device::DeviceType)

A matrix list ū, where ū[:,:,q] is a local basis rotation on qubit q.

CtrlVQE.Devices.localdriveoperators — Method

localdriveoperators(device[, basis], t; kwargs...)

A matrix list v̄, where v̄[:,:,q] represents a sum of all drives acting on qubit q.

Arguments

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the operators will be represented in. Defaults to Bases.OCCUPATION when omitted.

t::Real: the time each drive operator is evaluated at.

CtrlVQE.Devices.localdrivepropagators — Method

localdrivepropagators(device[, basis], τ, t; kwargs...)

A matrix list ū, where ū[:,:,q] is the propagator for a local drive term.

Arguments

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the operators will be represented in. Defaults to Bases.OCCUPATION when omitted.

τ::Real: the amount to move forward in time by. Note that the propagation is only approximate for time-dependent operators. The smaller τ is, the more accurate the approximation.

t::Real: the time each drive operator is evaluated at.

CtrlVQE.Devices.localloweringoperator — Method

localloweringoperator(device::DeviceType; result=nothing)

The lowering operator $a$ acting on the Hilbert space of a single physical qubit.

Optionally, pass a pre-allocated array of compatible type and shape as result.

CtrlVQE.Devices.localqubitevolvers — Method

localqubitevolvers(device[, basis], τ; kwargs...)

A matrix list ū, where each ū[:,:,q] is a propagator for a local qubit hamiltonian.

This function is identical to localqubitevolvers, except that the argument t is considered an absolute time so it is never cached.

Arguments

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the operator will be represented in. Defaults to Bases.OCCUPATION when omitted.

τ::Real: the amount to move forward in time by. Note that the propagation is only approximate for time-dependent operators. The smaller τ is, the more accurate the approximation.

Keyword Arguments

result: a pre-allocated array of compatible type and shape, used to store the result.

CtrlVQE.Devices.localqubitoperators — Method

localqubitoperators(device[, basis]; kwargs...)

A matrix list h̄, where each h̄[:,:,q] represents a local qubit hamiltonian.

Arguments

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the operators will be represented in. Defaults to Bases.OCCUPATION when omitted.

Keyword Arguments

result: a pre-allocated array of compatible type and shape, used to store the result.

localqubitoperators(device, basis, :cache)

For internal use only. The extra positional argument enables dispatch to a cached function when appropriate.

CtrlVQE.Devices.localqubitpropagators — Method

localqubitpropagators(device[, basis], τ; kwargs...)

A matrix list ū, where each ū[:,:,q] is a propagator for a local qubit hamiltonian.

Arguments

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the operator will be represented in. Defaults to Bases.OCCUPATION when omitted.

τ::Real: the amount to move forward in time by.

Keyword Arguments

result: a pre-allocated array of compatible type and shape, used to store the result.

localqubitpropagators(device, basis, τ, :cache)

For internal use only. The extra positional argument enables dispatch to a cached function when appropriate.

CtrlVQE.Devices.ndrives — Method

ndrives(device::DeviceType)

The number of distinct drive channels.

CtrlVQE.Devices.ngrades — Method

ngrades(device::DeviceType)

The number of distinct gradient operators.

CtrlVQE.Devices.nlevels — Method

nlevels(device::DeviceType)

The number of physical levels in each "qubit".

CtrlVQE.Devices.nqubits — Method

nqubits(device::DeviceType)

The number of qubits in the device.

CtrlVQE.Devices.nstates — Method

nstates(device::DeviceType)

The total number of states in the physical Hilbert space of the device.

(This is as opposed to nlevels(device), the number of states in the physical Hilbert space of a single independent qubit.)

CtrlVQE.Devices.operator — Method

operator(op, device[, basis]; kwargs...)

A Hermitian operator describing a device, represented in the given basis.

For example, to construct the static Hamiltonian of a device in the dressed basis, call operator(Operators.STATIC, device, Bases.DRESSED).

Arguments

op::Operators.OperatorType: which operator to construct (eg. static, drive, etc.).

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the operator will be represented in. Defaults to Bases.OCCUPATION when omitted.

Keyword Arguments

result: a pre-allocated array of compatible type and shape, used to store the result.

operator(op, device, basis, :cache)

For internal use only. The extra positional argument enables dispatch to a cached function when appropriate.

CtrlVQE.Devices.propagate! — Method

propagate!(op, device[, basis], τ, ψ)

Propagate a state ψ by a small time τ under the Hermitian op describing a device.

Arguments

op::Operators.OperatorType: which operator to evolve under (eg. static, drive, etc.).

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the state ψ is represented in. Defaults to Bases.OCCUPATION when omitted.

τ::Real: the amount to move forward in time by. Note that the propagation is only approximate for time-dependent operators. The smaller τ is, the more accurate the approximation.

ψ: Either a vector or a matrix, defined over the full Hilbert space of the device.

CtrlVQE.Devices.propagator — Method

propagator(op, device[, basis], τ; kwargs...)

A unitary propagator describing evolution under a Hermitian operator for a small time τ.

Arguments

op::Operators.OperatorType: which operator to evolve under (eg. static, drive, etc.).

device::DeviceType: which device is being described.

basis::Bases.BasisType: which basis the operator will be represented in. Defaults to Bases.OCCUPATION when omitted.

τ::Real: the amount to move forward in time by. Note that the propagation is only approximate for time-dependent operators. The smaller τ is, the more accurate the approximation.

Keyword Arguments

result: a pre-allocated array of compatible type and shape, used to store the result.

propagator(op, device, basis, τ, :cache)

For internal use only. The extra positional argument enables dispatch to a cached function when appropriate.

CtrlVQE.Devices.qubithamiltonian — Method

qubithamiltonian(device::DeviceType, ā::MatrixList, q::Int; result=nothing)

The static components of the device Hamiltonian local to qubit q.

This method is a function of annihilation operators $a_q$ given by ā[:,:,q], which may be matrices acting on a physical Hilbert space either globally or locally.

Optionally, pass a pre-allocated array of compatible type and shape as result.

CtrlVQE.Devices.staticcoupling — Method

staticcoupling(device::DeviceType, ā::MatrixList, q::Int; result=nothing)

The static components of the device Hamiltonian nonlocal to any one qubit.

This method is a function of annihilation operators $a_q$ given by ā[:,:,q], which are matrices acting globally on the physical Hilbert space.

Optionally, pass a pre-allocated array of compatible type and shape as result.

Transmon Devices

CtrlVQE.TransmonDevices.FixedFrequencyTransmonDevice — Type

FixedFrequencyTransmonDevice(ω̄, δ̄, ḡ, quples, q̄, ν̄, Ω̄, m)

A transmon device, modeling for example IBM's superconducting quantum computers.

Variational parameters include ONLY the shape parameters in each pulse. Pulse frequencies are "frozen".

Arguments

ω̄: a vector of angular frequencies specifying the resonance frequency of each qubit.

δ̄: a vector of angular frequencies specifying the anharmonicity of each qubit.

ḡ: a vector of angular frequencies specifying the strength of each coupling.

quples: a vector of Quple identifying whcih qubits participate in each coupling.

q̄: a vector of indices specifying the target qubit for each drive channel.

ν̄: a vector of angular frequencies specifying the pulse frequencies for each channel.

Ω̄: a vector of signals specifying the shape of the pulse for each channel.

m: an integer specifying the number of physical levels to retain for each qubit.

CtrlVQE.TransmonDevices.TransmonDevice — Type

TransmonDevice(ω̄, δ̄, ḡ, quples, q̄, ν̄, Ω̄, m)

A transmon device, modeling for example IBM's superconducting quantum computers.

Variational parameters include the shape parameters in each pulse, and the pulse frequencies.

Arguments

ω̄: a vector of angular frequencies specifying the resonance frequency of each qubit.

δ̄: a vector of angular frequencies specifying the anharmonicity of each qubit.

ḡ: a vector of angular frequencies specifying the strength of each coupling.

quples: a vector of Quple identifying whcih qubits participate in each coupling.

q̄: a vector of indices specifying the target qubit for each drive channel.

ν̄: a vector of angular frequencies specifying the pulse frequencies for each channel.

Ω̄: a vector of signals specifying the shape of the pulse for each channel.

m: an integer specifying the number of physical levels to retain for each qubit.